refactor : 제출페이지의 Network 탭에서 테스트 결과를 찾을 수없다는 이슈를 해결
