+++ src/test/java/lotto/ApplicationTest.java
@@ -56,14 +56,6 @@ class ApplicationTest extends NsTest {
         });
     }
 
-    @Test
-    void 예외_테스트_2() {
-        assertSimpleTest(() -> {
-            runException("12345");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
     @Override
     public void runMain() {
         Application.main(new String[]{});
